@@ -1,9 +1,14 @@
 module github.com/circonus-labs/go-trapcheck
 
-go 1.16
+go 1.17
 
 require (
-	github.com/circonus-labs/go-apiclient v0.7.15
+	github.com/circonus-labs/go-apiclient v0.7.18
 	github.com/google/uuid v1.3.0
-	github.com/hashicorp/go-retryablehttp v0.7.0
+	github.com/hashicorp/go-retryablehttp v0.7.1
+)
+
+require (
+	github.com/hashicorp/go-cleanhttp v0.5.1 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
 )
