@@ -8,18 +8,19 @@ package trapcheck
 import (
 	"crypto/tls"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"log"
 	"reflect"
 	"testing"
 
 	"github.com/circonus-labs/go-apiclient"
+	"github.com/circonus-labs/go-apiclient/config"
 )
 
 func TestTrapCheck_fetchCert(t *testing.T) {
 	tc := &TrapCheck{}
 	tc.Log = &LogWrapper{
-		Log:   log.New(ioutil.Discard, "", log.LstdFlags),
+		Log:   log.New(io.Discard, "", log.LstdFlags),
 		Debug: false,
 	}
 
@@ -89,7 +90,7 @@ func TestTrapCheck_fetchCert(t *testing.T) {
 func TestTrapCheck_setBrokerTLSConfig(t *testing.T) {
 	tc := &TrapCheck{}
 	tc.Log = &LogWrapper{
-		Log:   log.New(ioutil.Discard, "", log.LstdFlags),
+		Log:   log.New(io.Discard, "", log.LstdFlags),
 		Debug: false,
 	}
 
@@ -233,6 +234,9 @@ func TestTrapCheck_setBrokerTLSConfig(t *testing.T) {
 			tc.tlsConfig = tt.tlsConfig
 			tc.checkBundle = tt.checkBundle
 			tc.broker = tt.broker
+			if tc.checkBundle != nil {
+				tc.submissionURL = tt.checkBundle.Config[config.SubmissionURL]
+			}
 
 			if err := tc.setBrokerTLSConfig(); (err != nil) != tt.wantErr {
 				t.Errorf("TrapCheck.setBrokerTLSConfig() error = %v, wantErr %v", err, tt.wantErr)
